Moved BMIGraph.js to components/graph
