+++ frontend/src/components/graph/BMIGraph.js
